+++ demos/employees/src/main/java/is/codion/framework/demos/employees/ui/EmployeesAppPanel.java
@@ -52,7 +52,7 @@ public class EmployeesAppPanel extends EntityApplicationPanel<EmployeesAppModel>
     EntityPanel departmentPanel = new EntityPanel(departmentModel,
             new DepartmentEditPanel(departmentModel.editModel()),
             new DepartmentTablePanel(departmentModel.tableModel()),
-            settings -> settings.panelLayout(splitPaneResizeWeight(0.4)));
+            config -> config.panelLayout(splitPaneResizeWeight(0.4)));
     departmentPanel.addDetailPanel(employeePanel);
 
     return singletonList(departmentPanel);
@@ -73,7 +73,7 @@ public class EmployeesAppPanel extends EntityApplicationPanel<EmployeesAppModel>
     SwingEntityTableModel tableModel = SwingEntityTableModel.tableModel(entities, applicationModel().connectionProvider());
     tableModel.editModel().readOnly().set(true);
     EntityTablePanel tablePanel = new EntityTablePanel(tableModel,
-            settings -> settings.includePopupMenu(false));
+            config -> config.includePopupMenu(false));
 
     Dialogs.componentDialog(tablePanel.initialize())
             .owner(this)